+++ DREAM CATCH/YoutubeListViewController.swift
@@ -1,0 +1,175 @@
+//
+//  YoutubeViewController.swift
+//  DREAM CATCH
+//
+//  Created by 髙野拓弥 on 2021/04/28.
+//
+
+import UIKit
+import Alamofire
+
+class YoutubeListViewController: UIViewController {
+    @IBOutlet weak var profileimageView: UIImageView!
+    @IBOutlet weak var headerView: UIView!
+    @IBOutlet weak var headerHightConstraint: NSLayoutConstraint!
+    @IBOutlet weak var headerTopConstraint: NSLayoutConstraint!
+    
+    private var prevContentOfset: CGPoint = .init(x: 0, y: 0)
+    private let headerMoveHeight: CGFloat = 5
+    
+    @IBOutlet weak var videoListCollectionView: UICollectionView!
+    
+    private let cellId = "cellId"
+    private let attentionCellId = "atentionCellId"
+    private var videoItems = [Item]()
+    
+    override func viewDidLoad() {
+        super.viewDidLoad()
+        
+        setupViews()
+        fetchYoutubeSerachInfo()
+    }
+    
+    private func setupViews() {
+        videoListCollectionView.delegate = self
+        videoListCollectionView.dataSource = self
+        
+        videoListCollectionView.register(UINib(nibName: "VideoListCell", bundle: nil), forCellWithReuseIdentifier: cellId)
+        videoListCollectionView.register(AttentionCell.self, forCellWithReuseIdentifier: attentionCellId)
+        
+        profileimageView.layer.cornerRadius = 20
+    }
+    
+    private func fetchYoutubeSerachInfo() {
+        let params = ["q": "motivation"]
+        
+        API.shared.request(path: .search, params: params, type: Video.self) { (video) in
+            self.videoItems = video.items
+            let id = self.videoItems[0].snippet.channelId
+            self.fetchYoutubeChannelInfo(id: id)
+        }
+    }
+    
+    private func fetchYoutubeChannelInfo(id: String) {
+        let params = [
+            "id": id
+        ]
+        
+        API.shared.request(path: .channels, params: params, type: Channel.self) { (channel) in
+            self.videoItems.forEach { (item) in
+                item.channel = channel
+            }
+            
+            self.videoListCollectionView.reloadData()
+        }
+    }
+    
+    private func headerViewEndAnimation() {
+        if headerTopConstraint.constant < -headerHightConstraint.constant / 2 {
+            UIView.animate(withDuration: 0.2, delay: 0, usingSpringWithDamping: 0.9, initialSpringVelocity: 0.8, options: [], animations: {
+                
+                self.headerTopConstraint.constant = -self.headerHightConstraint.constant
+                self.headerView.alpha = 0
+                self.view.layoutIfNeeded()
+            })
+        } else {
+            UIView.animate(withDuration: 0.2, delay: 0, usingSpringWithDamping: 0.9, initialSpringVelocity: 0.8, options: [], animations: {
+                
+                self.headerTopConstraint.constant = 0
+                self.headerView.alpha = 1
+                self.view.layoutIfNeeded()
+            })
+        }
+    }
+}
+
+// MARK: - scrollViewのdelegateメソッド
+extension YoutubeListViewController {
+    // scrollViewがscrollした時に呼ばれるメソッド
+    func scrollViewDidScroll(_ scrollView: UIScrollView) {
+        headerAnimation(scrollView: scrollView)
+    }
+    
+    private func headerAnimation(scrollView: UIScrollView) {
+        DispatchQueue.main.asyncAfter(deadline: .now() + 0.5) {
+            self.prevContentOfset = scrollView.contentOffset
+        }
+        
+        guard let presentIndexPath = videoListCollectionView.indexPathForItem(at: scrollView.contentOffset) else { return }
+        if scrollView.contentOffset.y < 0 { return }
+        if presentIndexPath.row >= videoItems.count - 2 { return }
+        
+        let alphaRaitio = 1 / headerHightConstraint.constant
+        
+        if self.prevContentOfset.y < scrollView.contentOffset.y {
+            if headerTopConstraint.constant <= -headerHightConstraint.constant { return }
+            headerTopConstraint.constant -= headerMoveHeight
+            headerView.alpha -= alphaRaitio * headerMoveHeight
+        } else if self.prevContentOfset.y > scrollView.contentOffset.y {
+            if headerTopConstraint.constant >= 0 { return }
+            headerTopConstraint.constant += headerMoveHeight
+            headerView.alpha += alphaRaitio * headerMoveHeight
+        }
+    }
+    
+    // scrollViewのscrollがピタッと止まった時に呼ばれる
+    func scrollViewDidEndDragging(_ scrollView: UIScrollView, willDecelerate decelerate: Bool) {
+        if !decelerate {
+            headerViewEndAnimation()
+        }
+    }
+    
+    // scrollViewが止まった時に呼ばれる
+    func scrollViewDidEndDecelerating(_ scrollView: UIScrollView) {
+        headerViewEndAnimation()
+    }
+}
+
+// MARK: - UICollectionViewDelegate, UICollectionViewDataSource, UICollectionViewDelegateFlowLayout
+extension YoutubeListViewController: UICollectionViewDelegate, UICollectionViewDataSource, UICollectionViewDelegateFlowLayout {
+    
+    func collectionView(_ collectionView: UICollectionView, didSelectItemAt indexPath: IndexPath) {
+        print("didSelectItemAt")
+
+        let youtubeViewController = UIStoryboard(name: "Youtube", bundle: nil).instantiateViewController(identifier: "YoutubeViewController") as YoutubeViewController
+        youtubeViewController.modalPresentationStyle = .fullScreen
+        
+        self.present(youtubeViewController, animated: true, completion: nil)
+    }
+    
+    func collectionView(_ collectionView: UICollectionView, layout collectionViewLayout: UICollectionViewLayout, sizeForItemAt indexPath: IndexPath) -> CGSize {
+        let width = self.view.frame.width
+        
+        if indexPath.row == 2 {
+            return .init(width: width, height: 200)
+        } else {
+            return .init(width: width, height: width)
+        }
+    }
+    
+    func collectionView(_ collectionView: UICollectionView, numberOfItemsInSection section: Int) -> Int {
+        return videoItems.count + 1
+    }
+    
+    func collectionView(_ collectionView: UICollectionView, cellForItemAt indexPath: IndexPath) -> UICollectionViewCell {
+        if indexPath.row == 2 {
+            let cell = videoListCollectionView.dequeueReusableCell(withReuseIdentifier: attentionCellId, for: indexPath) as! AttentionCell
+            cell.videoItems = self.videoItems
+            
+            return cell
+            
+        } else {
+            let cell = videoListCollectionView.dequeueReusableCell(withReuseIdentifier: cellId, for: indexPath) as! VideoListCell
+            
+            if self.videoItems.count == 0 { return cell }
+            
+            if indexPath.row > 2 {
+                cell.videoItem = videoItems[indexPath.row - 1]
+            } else {
+                cell.videoItem = videoItems[indexPath.row]
+            }
+            
+            return cell
+        }
+    }
+}
